+++ python-package/xgboost/dask.py
@@ -60,14 +60,13 @@ from typing import (
 import numpy
 
 from . import collective, config
-from ._typing import _T, FeatureNames
+from ._typing import _T, FeatureNames, FeatureTypes
 from .callback import TrainingCallback
 from .compat import DataFrame, LazyLoader, concat, lazy_isinstance
 from .core import (
     Booster,
     DataIter,
     DMatrix,
-    FeatureTypes,
     Metric,
     Objective,
     QuantileDMatrix,
