+++ python-package/xgboost/sklearn.py
@@ -21,7 +21,7 @@ from typing import (
 import numpy as np
 from scipy.special import softmax
 
-from ._typing import ArrayLike, FeatureNames
+from ._typing import ArrayLike, FeatureNames, FeatureTypes
 from .callback import TrainingCallback
 
 # Do not use class names on scikit-learn directly.  Re-define the classes on
@@ -37,7 +37,6 @@ from .config import config_context
 from .core import (
     Booster,
     DMatrix,
-    FeatureTypes,
     Metric,
     QuantileDMatrix,
     XGBoostError,
